+++ C64/PRGArchive.h
@@ -32,7 +32,7 @@ private:
     uint8_t *data;
 		
     //! @brief   File size
-	int size;
+	size_t size;
 
     /*! @brief   File pointer
      *  @details An offset into the data array. 
@@ -79,8 +79,8 @@ public:
     const char *getTypeAsString() { return "PRG"; }
     
     bool fileIsValid(const char *filename);
-    bool readFromBuffer(const uint8_t *buffer, unsigned length);
-    unsigned writeToBuffer(uint8_t *buffer);
+    bool readFromBuffer(const uint8_t *buffer, size_t length);
+    size_t writeToBuffer(uint8_t *buffer);
     
     
     //
